refactor(DQ-55): add openTime, closeTime format

+++ src/main/java/api/docq/domain/review/repository/ReviewRepository.java
@@ -13,5 +13,5 @@ public interface ReviewRepository extends JpaRepository<Review, Long> {
     Page<Review> findAllByClinicId(@Param("clinicId") Long clinicId, Pageable pageable);
 
     @Query("SELECT AVG(r.starPoint) FROM Review r WHERE r.clinicId = :clinicId")
-    Double findAverageStarPointByClinicId(@Param("clinicId") Long clinicId);
+    Integer findAverageStarPointByClinicId(@Param("clinicId") Long clinicId);
 }
